+++ src/speech_mcp/ui/components/__init__.py
@@ -1,0 +1,17 @@
+"""
+UI components for the Speech UI.
+
+This package contains reusable UI components for the Speech UI.
+"""
+
+from speech_mcp.ui.components.audio_visualizer import AudioVisualizer
+from speech_mcp.ui.components.animated_button import AnimatedButton
+from speech_mcp.ui.components.tts_adapter import TTSAdapter
+from speech_mcp.ui.components.audio_processor_ui import AudioProcessorUI
+
+__all__ = [
+    'AudioVisualizer',
+    'AnimatedButton',
+    'TTSAdapter',
+    'AudioProcessorUI'
+]
