+++ src/speech_mcp/ui/pyqt/__init__.py
@@ -1,0 +1,12 @@
+"""
+PyQt UI implementation for the Speech UI.
+
+This package provides a PyQt-based UI for speech recognition and visualization.
+"""
+
+from speech_mcp.ui.pyqt.pyqt_ui import PyQtSpeechUI, run_ui
+
+__all__ = [
+    'PyQtSpeechUI',
+    'run_ui'
+]
